pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bereans-study-mcp"
version = "0.1.0"
description = "MCP server — a living knowledge base with semantic search"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "ebooklib>=0.18",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pandas>=2.2.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bereans"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
